优化逻辑，提升刷屏速度到10秒左右

+++ USB-Screen/USB-Screen/Dev/Hid.cs
@@ -30,13 +30,23 @@ namespace USB_Screen.Dev
         public bool SendBytes(byte[] data)
         {
             Progress = 0;//进度清零
+
+
+            HidDevice[] HidDeviceList;
+            HidDevice HidDevice;
+
+            // Enumerate the devices with the Vendor Id
+            HidDeviceList = HidDevices.Enumerate(0x2333, 0x2434).ToArray();
+            if (HidDeviceList.Length <= 0)
+                return false;
+
+            HidDevice = HidDeviceList[0];
             List<byte> d = new List<byte>(data);
             while (d.Count > 64)
             {
                 var temp = d.GetRange(0, 64);
                 d.RemoveRange(0,64);
-                var result = Send(temp.ToArray());
-                if (result.Length != 64)
+                if (!Send(temp.ToArray(), HidDevice))
                     return false;
                 Progress = 100 - (int)(100 * (double)d.Count / data.Length);//更新进度
             }
@@ -46,8 +56,7 @@ namespace USB_Screen.Dev
                 var temp = d.GetRange(0, d.Count);
                 for (int i= d.Count; i<64;i++)
                     temp.Add(0);
-                var result = Send(temp.ToArray());
-                if (result.Length != 64)
+                if (!Send(temp.ToArray(), HidDevice))
                     return false;
                 Progress = 100 - (int)(100 * (double)d.Count / data.Length);//更新进度
             }
@@ -61,53 +70,22 @@ namespace USB_Screen.Dev
         /// 发一包数据
         /// </summary>
         /// <param name="data">数据</param>
-        /// <returns>返回的数据</returns>
-        private static byte[] Send(byte [] data)
+        /// <returns>是否成功</returns>
+        private static bool Send(byte [] data, HidDevice HidDevice)
         {
-            Debug.WriteLine($"pack data {ByteToHexString(data)}");
             var temp = new byte[data.Length + 1];
             temp[0] = 0;
             for (int i = 1; i < data.Length + 1; i++)
                 temp[i] = data[i-1];
             try
             {
-                HidDevice[] HidDeviceList;
-                HidDevice HidDevice;
-
-                // Enumerate the devices with the Vendor Id
-                HidDeviceList = HidDevices.Enumerate(0x2333, 0x2434).ToArray();
-
-                if (HidDeviceList.Length > 0)
-                {
-                    // Grab the first device
-                    HidDevice = HidDeviceList[0];
-
-                    // Check if connected...
-                    //Debug.WriteLine("Connected: " + HidDevice.IsConnected.ToString());
-
-                    HidDevice.Write(temp);
-
-                    // Blocking read of report
-                    HidDeviceData InData;
-
-                    InData = HidDevice.Read();
-
-                    var id = InData.Data;
-
-                    var ti = new byte[id.Length - 1];
-                    for (int i = 1; i < id.Length - 1; i++)
-                        ti[i] = id[i + 1];
-                    return ti;
-                }
-                else
-                {
-                    return new byte[] { };
-                }
+                HidDevice.Write(temp);
+                return true;
             }
             catch(Exception e)
             {
                 Debug.WriteLine($"hid send error\r\n{e}");
-                return new byte[] { };
+                return false;
             }
         }
